+++ R/connect.R
@@ -67,7 +67,15 @@ Connect <- R6::R6Class(
     print = function(...) {
       cat("Posit Connect API Client: \n")
       cat("  Posit Connect Server: ", self$server, "\n", sep = "")
-      cat("  Posit Connect API Key: ", paste0(strrep("*", 11), substr(self$api_key, nchar(self$api_key) - 3, nchar(self$api_key))), "\n", sep = "")
+      cat(
+        "  Posit Connect API Key: ",
+        paste0(
+          strrep("*", 11),
+          substr(self$api_key, nchar(self$api_key) - 3, nchar(self$api_key))
+        ),
+        "\n",
+        sep = ""
+      )
       # TODO: something about API key... role... ?
       # TODO: point to docs on methods... how to see methods?
       cat("\n")
@@ -107,7 +115,13 @@ Connect <- R6::R6Class(
     #' @description Build a URL relative to the API root
     #' @param ... path segments
     api_url = function(...) {
-      paste(self$server, "__api__", ..., sep = "/")
+      self$server_url("__api__", ...)
+    },
+
+    #' @description Build a URL relative to the server root
+    #' @param ... path segments
+    server_url = function(...) {
+      paste(self$server, ..., sep = "/")
     },
 
     #' @description General wrapper around `httr` verbs
@@ -190,7 +204,7 @@ Connect <- R6::R6Class(
     #' will be returned. Otherwise, the argument is forwarded to
     #' `httr::content(res, as = parser)`.
     DELETE = function(path, ..., url = self$api_url(path), parser = NULL) {
-      self$request("DELETE", url, parser = NULL, ...)
+      self$request("DELETE", url, parser = parser, ...)
     },
 
     #' @description Perform an HTTP PATCH request of the named API path.
@@ -246,7 +260,7 @@ Connect <- R6::R6Class(
     #' @description Return all tags.
     #' @param use_cache Indicates that a cached set of tags is used.
     get_tags = function(use_cache = FALSE) {
-      error_if_less_than(self, "1.8.6")
+      error_if_less_than(self$version, "1.8.6")
       # TODO: check cache "age"?
       if (is.null(self$tags) || !use_cache) {
         self$tags <- self$tag()
@@ -299,7 +313,7 @@ Connect <- R6::R6Class(
     #' @param name The tag name.
     #' @param parent_id The parent identifier.
     tag_create = function(name, parent_id = NULL) {
-      error_if_less_than(self, "1.8.6")
+      error_if_less_than(self$version, "1.8.6")
       dat <- list(
         name = name
       )
@@ -322,7 +336,7 @@ Connect <- R6::R6Class(
     #' @description Get a tag.
     #' @param id The tag identifier.
     tag = function(id = NULL) {
-      error_if_less_than(self, "1.8.6")
+      error_if_less_than(self$version, "1.8.6")
       if (is.null(id)) {
         path <- v1_url("tags")
       } else {
@@ -352,7 +366,7 @@ Connect <- R6::R6Class(
         count = min(page_size, .limit)
       )
       if (!is.null(filter)) {
-        query$filter <- paste(sapply(1:length(filter), function(i) {
+        query$filter <- paste(sapply(seq_along(filter), function(i) {
           sprintf("%s:%s", names(filter)[i], filter[[i]])
         }), collapse = .collapse)
       }
@@ -818,6 +832,7 @@ Connect <- R6::R6Class(
 
     #' @description Get R installations.
     server_settings_r = function() {
+      lifecycle::deprecate_soft("0.3.1", "Connect$server_settings_r()", "get_runtimes()")
       self$GET(v1_url("server_settings", "r"))
     },
 
@@ -827,6 +842,18 @@ Connect <- R6::R6Class(
     }
 
     # end --------------------------------------------------------
+  ),
+  private = list(
+    .version = NULL
+  ),
+  active = list(
+    #' @field version The server version.
+    version = function() {
+      if (is.null(private$.version)) {
+        private$.version <- safe_server_version(self)
+      }
+      private$.version
+    }
   )
 )
 
@@ -880,7 +907,7 @@ connect <- function(
   tryCatch(
     {
       check_connect_license(con)
-      check_connect_version(using_version = safe_server_version(con))
+      warn_untested_connect(using_version = safe_server_version(con))
     },
     error = function(err) {
       if (.check_is_fatal) {