+++ tests/testthat/test-connect.R
@@ -73,7 +73,10 @@ with_mock_api({
     # Inject into this function something other than utils::browseURL
     # so we can assert that it is being called without actually trying to open a browser
     suppressMessages(trace("browse_url", where = connectapi::browse_solo, tracer = quote({
-      browseURL <- function(x) warning(paste("Opening", x))
+      browseURL <- # nolint: object_name_linter
+        function(x) {
+          warning(paste("Opening", x))
+        }
     }), at = 1, print = FALSE))
     expect_warning(
       browse_connect(con),
@@ -96,5 +99,19 @@ with_mock_api({
       ),
       "MY_MAGIC_HEADER: value"
     )
+  })
+
+
+  test_that("client$version is NA when server settings lacks version info", {
+    con <- Connect$new(server = "https://connect.example", api_key = "fake")
+    expect_message(v <- con$version, "Version information is not exposed by this Posit Connect instance")
+    expect_true(is.na(v))
+  })
+})
+
+test_that("client$version is returns version when server settings exposes it", {
+  with_mock_dir("2024.09.0", {
+    con <- Connect$new(server = "https://connect.example", api_key = "fake")
+    expect_equal(con$version, "2024.09.0")
   })
 })